requirements.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
cryptography>=41
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.2
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-markdownx==3.0.1
executing==1.2.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
ipython==8.6.0
jedi==0.18.2
Markdown==3.4.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.1
traitlets==5.5.0
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.7
wcwidth==0.2.5
zipp==3.10.0
